+++ src/index.scss
@@ -3,7 +3,6 @@
 @import "~@edx/paragon/scss/core/core";
 @import "~@edx/brand/paragon/overrides";
 
-@import "studio-header/header";
 @import "~@edx/frontend-component-footer/dist/footer";
 @import '~@edx/frontend-component-header/dist/index';
