+++ src/studio-header-wrapper/HeaderWrapper.test.jsx
